[dagit] Fix phantom overlay in playground
Summary:

On the Playground, the "loading" overlay is rendered whether anything is loading or not. My recent change to pointer-events now prevents clicks from making it through to the controls underneath.

Instead, just don't render the loading overlay if it's not loading.

Test Plan:

View Playground, verify that I can click on the "Launch Execution" button.

Reviewers:

bengotow, sashank, prha

Subscribers: