[Dagit] Fix frozen scrolling on short logs list
Summary:
Resolves #3331

Bug: when a logs list in the Run view is fairly short, the dynamic scrolling table might render the top of it beyond the top of the pane but not allow scrolling. This is because the container div is given 100px padding, which is applied after the measurement has occurred. The table doesn't think it needs a scrollbar, so it isn't given one, and the top rows are visually cut off and inaccessible.

Instead, apply the bottom padding to the last item in the list. This way, the measurement will taken it into account, and the scrollbar will be shown correctly.

Test Plan: View a run with a short log list. Verify that I am able to scroll to previously cut-off log entries.

Reviewers: bengotow, nate

Differential Revision: https://dagster.phacility.com/D5412