Countdown component
Summary:

Create a "Countdown" component, a storybook example, and a use case in `ScheduleRoot`.

`Countdown` is a controlled component that allows triggering the countdown via a prop (and rendering the desired contents), with the countdown interval itself managed within an effect. When the countdown completes, the time remaining will stay at zero, the parent is informed via `onComplete` and can set `status` to "idle" while necessary behavior is carried out. It can then re-trigger the countdown by again setting the `status` prop to "counting".

In the case of `ScheduleRoot`, the countdown is kicked off by changes in the network status of a polling GraphQL query. When the query changes to a "ready" state, it triggers the countdown. Any other network status indicates that fetching behavior is underway, so the Countdown should be in an "idle" state.

The `message` function allows the caller to render the countdown (or idle) data as it sees fit.

Test Plan:

View `Countdown` in storybook, use buttons to trigger the two states. Play around with `duration` and `interval`, verify desired behavior.

View Schedule in dagit, verify proper rendering of the countdown, verify click-to-refresh behavior.

Reviewers:

bengotow, prha

Subscribers: