[dagit] Clean up tables
Summary:

Clean up a few tables in Dagit. They have a tendency to overflow the window, in part because of the no-wrap behavior in the cells. I don't think we need this, so I'm removing it.

Test Plan:

View tables throughout Dagit, verify that they render better now.

Reviewers:

bengotow, sashank, prha

Subscribers: