Derive the schedule partition set directly from the execution times, to ensure they alwasy stay in sync
Summary:
Today there are two pieces of code in schedules that do very similar thing: there's an iterator that the scheduler uses to figure out when the ticks should be, and there's date_partition_range, which determines the partitions in the underlying partition set. This leads to subtle errors - for example, sometimes you'll skip a tick due to daylight savings time shenanigans, but that might not be reflected in the partition set since it's using different logic.

This diff tries to fix that by having the partition set generation function be derived from the same iterator that we use to generate the tick times.

Test Plan: BK (new test case that would have failed before)

Reviewers: sashank, prha