Don't say "No stack provided" for errors with no stack
Summary:
There are valid cases to show an error with no stack (e.g. a config issue). Drop the extra verbiage.

Test Plan:
Launch a backfill with a config error, bask in the glow of a nice clean error dialog
Launch a backfill with a system error, see the stack trace still

Reviewers: dish, bengotow