RFC: New mutation for force-mark-as-failed during termination
Summary:
Getting consensus on the API before writing tests - this adds a new argument that you can pass to termination to make it force-mark-as-failed. It will still try to terminate, but as long as the run isn't finished it won't require the termination to be possible (or wait for it to finish) before marking the run as failed. Some downstream changes will likely be needed to make sure that runs terminated in this way aren't making assumptions about the run status during execution.

Test Plan: Will add some BK tests based on feedback here

Reviewers: dish, johann, alangenfeld, catherinewu, prha

CC: schrockn