Disable catchup behavior for partitionless schedules
Summary:
Outcome from the failure meeting yesterday was to disable our auto-catchup behavior for schedules without partitions - the assumption is that if there's no partition set, it's probably the kind of schedule that doesn't actually need the catchup behavior (for example, a schedule that deletes and recreates a table) - in fact, in those cases, kicking off a bunch of runs at once could even be harmful.

Eventually we could also make this a flag on the schedule definition that you can configure?

Test Plan: New BK

Reviewers: sandyryza, sashank, schrockn

Differential Revision: https://dagster.phacility.com/D5480