[dagit] Disable Gaant for queued run
Summary:

Resolves #3364

When a run is queued, we shouldn't try to render a Gantt chart for it. Render a different state instead, with a link to view queued runs.

Test Plan:

Configure a queue coordinator but don't start the daemon, kick off a run. Verify that the Gantt placeholder renders.

Reviewers:

johann, dgibson, prha

Subscribers: