RFC DynamicOutput DSL __iter__ approach
Summary:
This demonstrates the iterator based approach for unpacking invoked solid dynamic outputs in the dsl. The updated test callsites give a flavor of how this plays out.

I think the clear next step if there is support for this approach is to also support a function like `map` for unpacking in addition to iterating.

Test Plan: added tests & updates

Reviewers: catherinewu, schrockn, sandyryza

Differential Revision: https://dagster.phacility.com/D5613