Add an HTML metadata entry type, allow JSON and HTML to be inlined into logs
Summary: This diff adds a new metadata entry type that provides a catch-all for users who want to display more interesting stuff in the log view. It was suggested in #3416 and seemed like it'd be useful, but I wouldn't object if we want to postpone this in favor of more codified graphing support later.

By default, HTML you provide is shown as `[Show HTML]` and the modal displays the content within a sandboxed iframe. If you pass `inline=True`, we trust you know what you're doing and allow the HTML to appear directly within the metadata entries table in the logs. This allows you to do cool stuff, but largely requires that your HTML snippet have a fully specified or immediately calculatable size and is fairly small. The screenshots below show it being used to embed an SVG with a bar chart, and a small image.

Test Plan: Run updated ManyEvents pipeline

Reviewers: @dish, @prha, schrockn

Subscribers: