Remove materializations / expectations state tracking from RunMetadataProvider
Summary: I think that this code dates back to when we had a visualization that showed a "summary" view of materializations and expectation results. We now show these in the structured logs and aren't actually using these fields on the summary stats!

Test Plan: See that everything still works without this code

Reviewers: @dish

Subscribers: