Eliminate pipeline_def from InitResourceContext
Summary: Preparing for a future where pipelines are no longer required
for things. I've encountered this both in the context of preparing the
graph/pipeline merge as well as making solids independently executable
in a first class way. I was pleasantly surprised to see that this did
not break any tests. The question is how defensive we want to be here.
Do we think this will break users?

Test Plan: integration

Reviewers: sandyryza, cdecarolis, yuhan, alangenfeld

Differential Revision: https://dagster.phacility.com/D5709