Fix dagster_github after pyjwt upgrade to 2.0
Summary:
This started failing after pyjwt upgraded to 2.0 due to not being able to call decode() on a string. I'm a little confused why this was passing before as .encode() seems to have returns a stirng and not bytes for a whilel.

Test Plan: run dagster-github tox (both with pyjwt pinned to the previous release and the new release)

Reviewers: alangenfeld, max, nate, schrockn