Remove DagstermillExecutionError altogether
Summary:
This used to be inside a user_error_code_boundary, and I left it in a weird state where it was kind of a user code error boundary and kind of not.

The thing is, this code is already covered by the step execution user error code boundary since it's inside a solid execution function, so we don't need another one. Can we just remove the redundant boundary altogether? I don't see anything catching this specific error, and it just gets discarded and replaced with its underlying PapermillExecutionError (the user code exception) when an exception is bubbled up.

Test Plan: BK

Reviewers: max, alangenfeld