Make DagsterDaemonScheduler the default scheduler
Summary:
Now that we have heartbeats on the daemons page I think it is reasonable to make this the default. We should think about what kind of affordances we might want to put into dagit that prompt people to go deploy the daemon though.

Reviewers: schrockn, alangenfeld, prha

Test Plan: BK, launch dagit with empty dagster.yaml