[dagit] Allow termination of entire backfill
Summary:

Resolves #3259

Allow bulk termination of a backfill from the partition set page.

When the view is filtered to a specific backfill tag that still has unfinished runs, a "Terminate" button is rendered. Clicking the button passes the list of unfinished runs to the termination dialog, which then attempts to terminate the runs on confirmation.

Test Plan:

Kick off a backfill. Verify that the Terminate button appears, click it. Verify correct termination workflow.

Reviewers:

sandyryza, catherinewu, bengotow

Subscribers: