[dagit] Filter out repositories with no schedules/sensors
Summary:

Resolves #3487

When showing schedule and sensor tables on the Daemons page, filter out repositories that don't have them.

This is better than rendering an empty table, and since it doesn't seem super valuable to have any empty state here after all, we might as well just not show anything for the empties.

Test Plan:

Force repo to behave as though it has no schedules/sensors, verify that nothing is rendered for it.

Reviewers:

prha, dgibson, johann

Subscribers: