[dagit] Create copy of DagsterGraphQLContext for each request
Summary: This diff updates the Dagit webserver such that the DagsterGraphQLContext is request scoped and clones the current workspace is at the top of the request. This is needed due to the introduction of the repository location watch threads that can modify the context. Now the watch threads only run on the long-lived global context.

Test Plan: unit

Reviewers: alangenfeld

Differential Revision: https://dagster.phacility.com/D5790