[dagit] Make sections of Instance config page linkable
Summary:

Enable linking to specific sections of yaml in the configuration section of "Instance details". We can use this to direct users to their configuration settings from elsewhere in Dagit, e.g. schedules.

For each top-level config section, there is a link that uses a hash to jump to that section. The hash-identified section will be highlighted in light gray. If no hash (or a missing hash) is used, nothing is highlighted.

Test Plan:

View `/instance/config`, verify correct rendering of yaml sections. Use link icon to highlight a section. Reload the page with that hash in place on the URL, verify that the browser jumps to it.

Reviewers:

johann, prha, dgibson

Subscribers: