migration for wiping the scheduler when transitioning from 0.9.x to 0.10.0
Summary: This makes it so that when you run dagster instance migrate after switching to 0.10.0, we drop the old schedule/schedule_ticks table (and wipe the scheduler to clear out any hanging cron jobs). There is not currently anything that triggers an InstanceMigrationRequired exception though.

Test Plan: BK (Still writing the postgres one)

Reviewers: prha, sashank, alangenfeld

Differential Revision: https://dagster.phacility.com/D5923