RFC: Make default sharded sqlite implementation asset-aware
Summary:
Was going to try to do a migration of the default sqlite storage to the
ConsolidatedSqliteEventLogStorage, but then thought about this approach, which seems like it might
work and is way less complicated.

It essentially mirrors all asset materialization events into a central assets.db sqlite file and
performs asset queries off of that db.

Test Plan: bk

Reviewers:

Subscribers: