Don't over-promise about daemon latency in the k8s docs
Summary: We don't currently offer 100ms latency, particularly if the scheduler happens to be doing a loop at the same time that runs are being taken off the queue. While this is a good goal, we shouldn't promise it in the docs until we can guarantee it.

Test Plan: View k8s docs.

Reviewers: johann, nate, catherinewu

Differential Revision: https://dagster.phacility.com/D6015