Flip asset materialization UI so newest time/partition is on the left
Summary: This diff flips the asset materialization UI so it reads newest-to-oldest.

Because we have sparklines beside each row, I flipped them to align with the data being presented in the grid. Since it seemed like the sparklines and the expanded graphs beneath should also match, I flipped those graphs as well so that the newest data is on the left. I think this is pretty nice overall and we could consider flipping other graphs in the app as well.

Test Plan: Run tests

Reviewers:

Subscribers: