Make type_check_fn optional on DagsterType
Summary:

I came across this when trying to write a guide on how to smoke test a data pipeline.  The smoke test uses dagster types to enable IO managers to auto-generate input data.  There's a lot going on, so I wanted to build it up incrementally, starting with the bare minimum that's required to get the example working.  The `type_check_fn` forces a digression.

More generally, there are a few ways that types offer value without a type_check_fn:
* Documenting outputs in Dagit
* dagster_type_loaders

Test Plan: bk

Reviewers:

Subscribers: