adds future schedule tick preview options
Summary: We want to be able to preview tags, run keys, and config of upcoming schedule ticks.

This diff lazily loads the run requests for the tick after a click into the actions menu, and either
shows direct actions (in the case of a single run request), or opens a dialog with multiple rows
of run requests inspection options.

Test Plan: Viewed upcoming ticks, viewd config, opened in playground

Reviewers: dgibson, dish

Subscribers: