Use IO managers for Dagstermill serialization
Summary: This replaces the old bespoke Dagstermill serialization scheme (which used `scrapbook` and could not handle inputs and outputs to/from Dagstermill notebooks that could not be JSON-serialized) with the IO managers.

Inputs are read directly by code invoked in the Dagstermill Jupyter process, and no longer cross the process boundary, but the `handle_input` method on the IO manager handling them must be idempotent, as this code will be invoked twice (once by the framework as it prepares to execute the Dagstermill solid, and once in the Jupyter process).

Outputs are persisted by calling `handle_output` within the Dagstermill Jupyter process, then by calling `handle_input` within the Dagstermill solid compute function, which finally yields them in the ordinary way back to the framework (which will then call `handle_output` again). The key assumptions here are that `handle_output` is idempotent and that that `handle_input(handle_output(X)) == X` for any value `X` that is yielded from a Dagstermill notebook and the IO manager on that value.

I don't believe there's a clear way to enforce or check these assumptions on these IO managers, but writing this has made me think that we should strongly discourage asymmetric IO managers in general, as they're difficult to reason about.

This also uses the IO managers to handle the output notebook (which is yielded as a simple bytes object), rather than the file manager, which is no longer provided by system storage machinery. There is a regression in functionality here, since the output notebook is no longer automatically available as a file on some file system, and no asset materialization is emitted, but in the absence of clear patterns for how to handle files/how materializations should be generated I think this is the safest way to proceed and involves the smallest risk of future user-facing breaking changes. Specifically, I think continuing to rely on the file manager API (which is a vestigial resource ABC that is never used by framework code) is misguided unless we are willing to stand behind it as the blessed approach for interacting with filesystems (in which case the framework should provide it). However, in order to restore the previous behavior, users will have to write and specify a custom IO manager on the output notebook outputs.

Like the inadvertent change documented in D6052, this is a breaking change because we are now requiring a different set of resources (`{"io_manager"}` instead of `{"file_manager"}`), and there is no way to specify optional resources for Dagster solids or to determine which resources will be available at solid construction time. This diff offers backcompatibility by retaining the old behavior of `define_dagstermill_solid` and introducing a new API with the new breaking behavior, `dagstermill_solid`.

I'd specifically like feedback on:

- The breaking change/deprecation scheme (vs. simply breaking users again as we inadvertently did in 0.10.0)
- The choice to move the output notebook to use the IO managers rather than retaining the file manager requirement

Test Plan: Unit

Reviewers: sandyryza, yuhan, alangenfeld, prha

Subscribers: