Replace dagster api grpc with 'python -m dagster.grpc' in docs?
Summary:
From digging into https://stackoverflow.com/questions/65860584/how-to-start-the-grpc-server-when-definitions-are-in-a-module/65863119#65863119 - it seems like when you run 'dagster api grpc', it has you in a weird state where your working directory(os.cwd) is the local directory, but the local directory is not in your sys.path???? But running python -m dagster.grpc with the same arguments does not have this problem (sys.path[0] is the folder from which you ran the command).

That confuses me, I guess this diff is really just a 'does anybody understand this'? If that's expected, we should probably change the docs to use python -m dagster.grpc instead of dagster api grpc... but I'm pretty sure that's not right, because that would also affect APIs like 'dagster pipeline execute'.

Maybe what this means is that we should just finish the job and let you specify working directories on modules like you can on files.

Test Plan: Not sure yet

Reviewers: prha, alangenfeld