Pseudocode RFC: Communicate via dagster host cloud and dagster user cloud via daemon message bus?
Summary:
This is a direction that both johann and alex proposed independently for getting repository information (in a user cloud) up into dagit (in a hosted cloud) without dagit needing to communciate directly with the daemon (or vice versa). It seemed crazy at first, but might be manageable if we don't mind a bit of latency? Just wanted to try out a very rough cut to see what it felt like in practice.

Obviously extremely rough/pseudo-code-y/no intention to land right now, and there's no implementation yet of any storage on the dagster cloud side (either the DB for storing metadata, or the queue for passing requests and responses back and forth).

The idea is:
- Users register a repository location with dagster cloud - similar to a gRPC workspace entry today, it includes a host, port, and location name that the daemon in the user cloud can use to reach it. They spin up a server locally with the repository info. At this point we also serialize the ExternalRepositoryData for each repository and put it in a DB, since it won't change from run to run and dagit needs it every time it spins up.

- When dagit spins up, it load the repostiory info from storage. Most of the time, this is enough for dagit to do its thing. When you do need to make an API call (for example, fetching partition information), you can write a request to a queue from host cloud, and a daemon in user cloud will read that request and write in a response - the only direct communication via host and user would be through this queue. We'd need to try this out to see how terrible the latency would be in practice here.

- When you go to enqueue a run, it works pretty much exactly how it does today - the run goes into the DB queued, and the daemon (over in the user cloud) reads it back out and launches it.

- When you update code, you also run a command to persist the latest versions of the ExternalRepositoryData.

Test Plan: no

Reviewers: max, johann, alangenfeld, schrockn, nate, catherinewu

Subscribers: