[dagit] RFC: Launch runs in the same tab
Summary:

From user feedback in Slack:

```
Is there any way to prevent dagit from opening new windows every-time that I execute a pipeline? While testing it's a bit annoying having to close windows. Also, the behaviour is not consistent anyway, if you use the "re-execute All (*)" in the same window it doesn't open a new window...
```

It looks like this behavior is intentional, but I figured I'd put up a diff to discuss making the behavior consistent by always opening the launched run in the same tab.

Test Plan:

Launch runs, verify that they are launched in the current tab.

Reviewers:

bengotow

Subscribers: