Update the headers on the run table to call out Run ID vs Snapshot ID, pipeline more clearly
Summary: I flipped the status and runID columns so that the Run ID column could have a header and not run into the "Actions" menu. I also replaced "Execution Params" with "Mode" because that column specifically displays the Mode and nothing else. I think we could potentially fold the Mode into the Pipeline area, but I want to avoid making it look like a tag.

Test Plan: Run tests

Reviewers: @dish, @schrockn

Subscribers: