Use a finite spinner for the partition view loading indicator
Summary: This replaces the "Loading Partitions..." spinner with one that fills from 0->100% as the data is loaded so it's less confusing when you view "all" parittions.

This also fixes a small bug in the url encoding of the query string so that you can reload the page when viewing "all" and still get "all".

Test Plan: Run tests

Reviewers: schrockn, dish

Subscribers: