[execute_in_process improvements 1/n]Pass value dict to graph during execution
Summary: This enables execute_in_process to take a dictionary of values to plug in to the top level graph. Note that this does not enable passing inputs under composition. We're consciously making the decision that inputs should be piped through the top level graph in order to support that use case.

Test Plan: Unit tests

Reviewers: alangenfeld, sandyryza, nate

Subscribers: