Repair pointer-events with Group margins
Summary:

I ran into an issue with `Group` yesterday. When the negative margin of a nested Group overlaps interactive contents of its container Group, those contents cannot be clicked.

We need to set `pointer-events` accordingly. The wrapper of the Group should have `none`, and the children should be set back to `auto`. This allows the mouse to reach the underlying interactive contents while keeping the children interactive as normal.

I created a storybook story to demonstrate.

Test Plan:

View "Pointer events test" story, verify that rendering and behavior are correct.

Reviewers:

bengotow

Subscribers: