repository decorator mypy
Summary: Adds type annotations for repository decorator. A repository definition is possibly returned by decorator, and by default a repository definition is not callable. The update_wrapper fxn makes it callable (this is my interpretation of what is happening), but mypy isn't smart enough to figure that out and I can't think of a way to let it know. To solve this, I added an unimplemented __call__ method on the repository definition class. This should always be overridden by the decorator.

Test Plan: Unit tests and mypy

Reviewers: alangenfeld

Differential Revision: https://dagster.phacility.com/D6338