[dagit] Preserve solid selection when refreshing config
Summary:

Based on a user report. With the change to allow refreshing configs after a manual repo reload, the user's current solid selection is not preserved.

Fix this by keeping the `currentSession.solidSelection` around, if any.

Test Plan:

View pipelines with presets and partition sets. Use a solid selection, then reload the repo and refresh config. Verify that the solid selection is preserved.

Reviewers:

bengotow, catherinewu, sashank

Subscribers: