[dagit] Revert code splitting for ExecutionSessionContainer
Summary:

Resolves #3696

Unfortuantely, I need to roll back the code splitting I put in place for ExecutionSessionContainer, which allowed us to get CodeMirror out of the main bundle. Because the Dagit web app files are prebuilt, the CRA has no awareness of the end user's indended path prefix, and all static file paths are set. Our `app.py` does some hackery to put the intended path prefix in place on the build `index.html`, but this will only apply to the main built files, not to lazy-loaded files like this one.

The result is that the lazy-loaded files can't be found at the default `/static/...` path, and the page error-crashes.

I believe our custom monospace font is failing to load for these users for the same reason.

Unless we can come up with a way to get the path prefix in place on the built files or some other clever solution, I think we're stuck with the non-split build. A simple revert here will unblock users who are trying to upgrade.

Test Plan:

```
make rebuild_dagit
dagit -p 3333 --path-prefix=/asdf
```

Open the browser, verify that I can navigate to the playground successfully.

Reviewers:

alangenfeld, bengotow, dgibson

Subscribers: