[dagit] Flip hideNonMatches default
Summary:

Relates to #3688.

In the Run log viewer, `hideNonMatches` currently defaults to false, so non-matching rows will appear in context alongside filter-matching rows.

I'm going to flip this default to `true` instead, as it seems like it might be a point of confusion that filtering doesn't always clearly surface results.

The UI is otherwise unchanged.

Test Plan:

View Run, do some filtering. Verify that the "Hide non-matches" box is unchecked by default. Check it, verify that non-matching logs are shown and that the URL updates accordingly.

Reviewers:

schrockn, bengotow

Subscribers: