Message missing repository errors in the scheduler differently in dagit
Summary:
Two small things to better handle the case in the scheduler when you rename a repo while a schedule is running:

- Include a recommendation of what to do (go turn off the now-unloadable schedule)
- Don't recommend restarting the daemon just because it raised an error - we could show something like this when the daemon is *down*, but if it's running and just raised an error we shouldn't recommend a restart.

(When the issue was filed the daemon was *also* actually crashing due an unrelated issue with the liveness check, johann fixed that separately)

One thing I'm not sure about is if we should message this as a 'health check'-type error at all? Since the daemon is running just fine, it's just being given data that's no longer valid so it can't run a schedule. i.e. its more of a user code error than an ops/admin error.

Reviewers: prha, schrockn, alangenfeld

CC: dish

Test Plan: BK, rename a repository witha  running daemon and dagit, dialog now includes a thing you should go do