Throw if you run dagit without DAGSTER_HOME set
Summary:
Particularly once the daemon is in the mix, letting you move forward in dagit with an in-memory DagsterInstance is just asking for trouble. Mimic the daemon behavior and throw until you set the DAGSTER_HOME environment variable.

The downside is a tiny bit more friction for the initial adoption of dagit - we might need to tweak the tutorial as well to include setting DAGSTER_HOME. You could probably convince me to make this just a warning.

Test Plan: Integration, launch dagit with and without DAGSTER_HOME set

Reviewers: sashank, alangenfeld, schrockn

Differential Revision: https://dagster.phacility.com/D6609