fix backfill status submitted count
Summary:
This is a short-term fix for displaying the number of runs still to be submitted.  The long
term fix will be a link to a backfill page, with more detailed information.  Previously, we were
showing that there were no submitted runs for all in-progress backfill states

Test Plan: bk

Reviewers: dgibson, sandyryza

Subscribers: