[dagster-postgres] cleaning up ScheduleStorage suite
Summary: I noticed that the fixture `clean_storage` was only used in one place so did a bit of housekeeping to make the code simpler

This diff exists for the same reason as D6714 - namely @prha's comments in D6617 about the MySQL schedule_storage tests (e.g. why the fixture if it's only used once).

Test Plan: bk

Reviewers: @bob, @catherinewu, @rexledesma, @prha

Subscribers: