Don't keep the grpc server process
Summary:
We originally added this lazy-load-user-code parameter because it was the only way to get error information surfaced in dagit. Now our loading path is better set up to handle errors (and surface them), so there's no need for this parameter, we can let the process die right away and surface the error in dagit (rather than waiting for the list_repositories error to fail).

Test Plan: BK, load dagit pointing at a file with a syntax error, syntax error still surfaced in dagit