Add 'use_ssl' boolean to grpc server workspace entry
Summary:
If this boolean is set, we use a secure channel in the dagster gRPC client.

Future work could include:
- letting you specify a .pem file with the specific credentials for the client to use
- setting up the server to accept SSL

The user request that triggered this does not require either of these, (The SSL routing is happening before it reaches the gRPC server, so the server still receives an insecure connection).

Test Plan: BK, will send this to the user as well to verify that it meets their requirements

Reviewers: alangenfeld, sashank, catherinewu

Differential Revision: https://dagster.phacility.com/D6743