[dagit] Show timezone in scheduled ticks table
Summary:

The timestamp column in Scheduled ticks is confusing, since it uses the schedule's execution timezone but doesn't display it. We should show the timezone here.

I probably broke this with my timestamp change.

Test Plan:

View Scheduled tab in `/instance/runs`, verify that timezone is shown with the timestamp.

Reviewers:

prha, johann, dgibson

Subscribers: