Always wait for grpc subprocesses to exit when daemon is terminated
Summary:
Test failures suggest that even if we don't wait for the subprocesses to finish cleanly, the daemon process won't abort cleanly and will just hang. So instead of making this a test-only flag, always have the daemon wait for any gRPC processes that it spun up.

Test Plan: Integration, BK

Reviewers: johann, prha, alangenfeld