[dagit] Allow Group wrapping
Summary:

This is a visual bug in Group that I've had on my hitlist for a while, going to go ahead and take care of it now. (No need to jam this in for this week's release.)

`Group` currently doesn't allow flex-wrapping of its children, so we can end up with some awkward spacing. This is especially noticeable on `TimestampDisplay`, which has the timestmap string adjacent to a clock icon. When space is insufficient, the timestamp text wraps and the icon stays where it is, resulting in kind of a crummy looking component.

By allowing `wrap`, the clock icon can bump down and keep the spacing a little more coherent. See screenshots.

Test Plan:

View Schedules table, shrink viewport width. Verify wrapping behavior of timestamp component.

Reviewers:

bengotow

Subscribers: