add instance arg to `build_resources` to ensure fs_io_manager can be initialized.
Summary: The fs_io_manager initialization process requires a DagsterInstance to get the default directory. Initializing IO managers is one of the most important functions of the `build_resources` API, so this adds an instance arg and confirms that the fs_io_manager can be instantiated and used to retrieve inputs.

Test Plan:
Added additional test that initializes fs_io_manager.
integration

Reviewers: alangenfeld

Differential Revision: https://dagster.phacility.com/D6808