[core][dagster-mysql] Fixing @experimental tag to work for classes
Summary: I was inspired by D6751, but I noticed that since the `@experimental` decorator treats its arguments as functions, the decorator & classes that rely on the ConfigurableClass mixin were incompatible. I added a custom decorator for classes. I didn't utilize subclassing since we want the class' signature (all of the metadata to be associated with it) to be the same with or without the decorator - I think with subclassing we'd need a lot more boilerplate (i.e. to overwrite attributes like __name__, etc.)

Test Plan: bk & local testing w/ dagit (via loading from dagster.yaml)

Reviewers: rexledesma, sandyryza

Differential Revision: https://dagster.phacility.com/D6810