Remove ManagedGrpcPythonEnvRepositoryLocationHandle, share more grpc server logic between dagit and the daemon
Summary:
Instead of having one path for loading gRPC servers in dagit and a totally different one in the daemon, consolidate them. The Workspace now also uses a RepositoryLocationHandleManager and a ProcessGrpcServerRegistry to manage handle lifecycles, just like the daemon.

This also opens up a future where we can pass in the handle manager instead of the ExternalPipeline to RunLauncher.launch_run, which will be a big step to being able to resolve https://github.com/dagster-io/dagster/issues/3778.

Test Plan: Integration

End goal here is to consolidate code for loading managed gRPC processes between the daemon and dagit / all dagster processes, with the goal of making so that you can pass in a handlemanager to RunLauncher rather than an ExternalPipeline. This will allow most run launchers to execute without ever having to load the ExternalPipeline.