Add snapshotId support to PiplineAndMode, display mode in more places
Summary:
This diff refactors PiplineAndMode a bit so that you can optionally include the snapshotId, which is rendered after the pipeline name and before the mode (per discussion with Alex).

I also expanded the places PipelineAndMode is used so that the mode is visible in a few more places, and renamed it PipelineReference which I think makes more sense now that renders the full specification of name+snapshot+mode with linking.

I also gave the component more ownership of the link generation, because I think asking people to provide a RepoAddress or explicitly type "repo-unknown" will encourage use of the RepoAddress whenever it's available, and also make sure you have to explicitly opt-out of linking anywhere the pipleine is referenced.

Test Plan: Review screenshots

Reviewers: dish

Differential Revision: https://dagster.phacility.com/D6824