longitudinal pipeline reframing
Summary:

I wanted to concrete-ize what the longitudinal is (pretend) doing so that:
* In screenshots, it'll be more obvious what's going on.
* It's clear what kind of metadata to add to AssetMaterializations.
* It's more straightforward to talk about in demos.

This reframes the longitudinal pipeline as a pipeline that's building tables that are used to train
a video recommendation model.

prha - if you'd prefer to keep the old framing that's fine too and maybe we can chat about how to
flesh it out.

Test Plan: ran it, bk

Reviewers: owen, prha

Subscribers: