[dagit] Don't show button group for non-partition assets
Summary:

For non-partition assets, don't show the "By Timestamp | By Partition" button group since it's not necessary.

Also remove the breadcrumbs in the page `Heading` since the breadcrumbs are visible in the line beneath it.

Test Plan:

View assets in Dagit, verify that partition assets have the button group and that non-partition assets do not.

Reviewers:

sandyryza, prha

Subscribers: