Allow for instantiated resources to be passed in to build_resources
Summary: This diff allows people to provide direct values as resources to build_resources. This is useful when trying to build a particular resource that might require other resource keys, and you don't want to necessarily provide the "real" versions of those resources.

Test Plan: Added test for instantiated resources and io managers

Reviewers: alangenfeld, sandyryza

Subscribers: